requirements.txt
celery
cherrypy
cloudevents-python
jsonpath2
pacifica-cli
pacifica-dispatcher
pacifica-downloader
pacifica-uploader
peewee
setuptools
six
git+https://github.com/IMMM-SFA/proxymod.git
